[WIP] Switch to an alpine base image
Consider a Meteor project that uses a native C++ module.
For example `bcrypt`. Since we are building the dependencies
here in the Ubuntu base image, copying the resulting binary
over to the Alpine based runtime image will not work properly,
likely leading to segmentation faults due to the glibc vs. musl c
mismatch.

This commit attempts to fix that, however it doesn't actually
work since the image fails to build on the very last step, where
the `meteor` command is attempting to run its own bundled binary
of `node` rather than the one available in the `$PATH`. This similarly
fails due to the glib vs. musl c mismatch.

So I feel like this is a step in the right direction, however I'm
not yet sure how to get past the fact that `meteor` wants to invoke
its (incompatible) bundled version of `node` instead of the system one.